+++ lib/mongoid/fields.rb
@@ -55,6 +55,8 @@ module Mongoid #:nodoc
 
       alias :id :_id
       alias :id= :_id=
+
+      attr_protected :id, :_id, :_type
     end
 
     # Apply all default values to the document which are not procs.